workfinder-master/requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.10.0
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
Django==4.1.2
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
iniconfig==1.1.1
isort==5.10.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
